moulitest bug fix, all non-bonus tests pass

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: cnysten <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/31 20:57:20 by cnysten           #+#    #+#             */
-/*   Updated: 2022/02/03 16:26:06 by carlnysten       ###   ########.fr       */
+/*   Updated: 2022/02/11 23:06:51 by carlnysten       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ int		ft_islower(int c);
 int		ft_isupper(int c);
 int		ft_atoi(const char *str);
 void	ft_bzero(void *s, size_t n);
-size_t	ft_intlen(int value);
+size_t	ft_intlen(long value);
 size_t	ft_intlen_base(int value, int base);
 int		ft_isalnum(int c);
 int		ft_isalpha(int c);
